Configure rustfmt for readability
The default configuration for the rustfmt tool includes settings
for brace style and line length that is not as readable as can be.

Regarding brace style, the K&R brace style is used as the default which,
in some circumstances (eg confusion where a function argument/parameter
list ends and the function's code block begins) makes reading and/or
understanding the source harder than it should be. This leads to a much
higher chance of errors being introduced when updating or adding to
code.

Regarding line length, 100 characters is the default line length. With
this default length, the source code is not as easy to read given
different displays, and as a result slows down understanding. Also,
studies have demonstrated some evidence that understanding of written
words is best achieved with arranged in a columnar format (eg
newspapers) as opposed stretching across the screen; with a line length
of 100 characters, source code no longer appears in columnar format.
These issues combined make the source code harder to read especially
when factoring different displays (eg online source code repositories
such as github, git command line tools such as git diff, different
devices ie desktop monitor vs phone or tablet).

This change adds a rustfmt configuration file that enforces Allman brace
style as well as enforcing a maximum line length of 80 characters. The
intent of these changes is to make the source code readable as on as
wide a number of displays as possible.