Add top-level tmp directory to .gitignore
Sometimes during development, there is a need to create temporary files
for various purposes (eg experimentation, temporarily saving a version
of a specific file, create randome data files, etc). However, without a
place inside the project filesystem hierarchy to do so, the chance of
adding such temporary files to the repository by accident is high.

This change adds the top-level tmp/ directory to the .gitignore file
which ensures no files contained in the directory will ever be added to
the repository.