cargo.toml: add new dev dependency: mocktopus
As this project is meant to traverse an entire directory tree, path
handling will be needed, as well as path related functions such as
getting the current directory. However, a lot of that functionality
resides in external code (eg in rust's standard library) and actually
affects the system (eg changing a directory). This poses difficulties
when trying to write unit tests since the default test runner runs tests
in separate threads.

To take getting the current directory and changing directory as
examples, one test may depend on changing the directory to a specific
location, and a second separate test may need to do the same thing but
changing to a different directory. However, depending on the order that
these 2 tests start and how fast they execute in their respective
threads, one or both of these tests may always fail when run together,
but always pass when only one of the tests is run at a time.

To solve this, external code will need to be mocked in a thread-local
manner. To achieve this, a new dev dependency on the mocktopus crate is
added to Cargo.toml which grants exactly this ability to create
thread-local mocks.

Please note: due to the nature of the rust programming language,
external code cannot be mocked directly. However, such code can always
be wrapped in custom functions or structs, and it is this custom code
that is able to be mocked.