Add standard comments to main.rs
The main.rs file was auto-generated by the cargo new command. As such,
it is missing helpful comment separators that group sections of the file
into specific purposes. This makes it easier to mess up the source code
when updating or adding something new, leading to code that is harder to
read and thus harder to undertand.

This change adds comments to the main.rs file that clearly shows the
purpose of each section of the file. For example, there are sections
for importing crates, defining modules, and the main entrypoint. By
doing so, this file adheres to a standard layout that all other files in
the project use, and it also makes it easier to read and understand.